@@ -6,18 +6,17 @@
     <parent>
         <artifactId>keycloak-parent</artifactId>
         <groupId>org.keycloak</groupId>
-        <version>18.0.2</version>
+        <version>22.0.4</version>
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
     <groupId>dev.suvera.keycloak</groupId>
     <artifactId>scim2-storage</artifactId>
-    <version>18.0.2</version>
     <packaging>jar</packaging>
     <description>Keycloak SCIM2 Outbound User Provisioning</description>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
 
         <guava.version>29.0-jre</guava.version>
         <commons.text.version>1.8</commons.text.version>
@@ -33,8 +32,8 @@
         <maven.compiler.target>1.8</maven.compiler.target>
 
         <json-schema-validator.version>2.2.14</json-schema-validator.version>
-        <validation-api.version>2.0.1.Final</validation-api.version>
-        <bean.validator.version>6.1.6.Final</bean.validator.version>
+        <validation-api.version>3.0.2</validation-api.version>
+        <bean.validator.version>7.0.1.Final</bean.validator.version>
         <!-- <log4j.version>2.13.3</log4j.version> -->
     </properties>
 
@@ -60,8 +59,8 @@
         </dependency>
 
         <dependency>
-            <groupId>javax.validation</groupId>
-            <artifactId>validation-api</artifactId>
+            <groupId>jakarta.validation</groupId>
+            <artifactId>jakarta.validation-api</artifactId>
             <version>${validation-api.version}</version>
         </dependency>
 
