@@ -8,8 +8,8 @@ import dev.suvera.scim2.schema.enums.ReturnedType;
 import dev.suvera.scim2.schema.enums.UniquenessType;
 import lombok.Data;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.Set;
