@@ -4,8 +4,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import dev.suvera.scim2.schema.data.ExtendedRecord;
 import lombok.Data;
 
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.PositiveOrZero;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.PositiveOrZero;
 import java.util.List;
 import java.util.Set;
 
