@@ -74,7 +74,7 @@ public class SkssStorageProviderFactory implements UserStorageProviderFactory<Sk
                 .add()
 
                 .property()
-                .label("Version: " + SkssStorageProviderFactory.class.getPackage().getImplementationVersion() + "-15")
+                .label("Version: " + SkssStorageProviderFactory.class.getPackage().getImplementationVersion() + "-15-buyah")
                 .helpText("Plugin version")
                 .add()
 
